+++ Changelog.md
@@ -52,6 +52,7 @@
 - Change: The Universal Key Changer can now be used by players in creative mode on any passcode-protected block or briefcase, regardless of its owner
 - Change: Passcodes no longer have a maximum length restriction
 - Change: Initially setting the passcode of a Briefcase now uses the same UI as inserting the passcode when opening a Briefcase
+- Change: Reinforced wool and carpets can now be dyed using any color wool or carpet
 - API: IModuleInventory#getModuleDescriptionId to make it possible to have shared descriptions
 - API: Renamed Option#readFromNBT to Option#load and Option#writeToNBT to Option#save
 - API: All occurrences of "password" in API class and method names have been switched out with "passcode" (for example, IPasswordProtected#getPassword -> IPasscodeProtected#getPasscode)